sshd-regen-keys: fix race condition
Systemd waits with starting service until a oneshot is finished this leads
to a race condition if you try to restart a service in a oneshot.

"Behavior of oneshot is similar to simple; however, the service manager will consider
the unit started after the main process exits. It will then start follow-up units.
RemainAfterExit= is particularly useful for this type of service.  Type=oneshot is the
implied default if neither Type= nor ExecStart= are specified."[1]

[1]: man systemd.service

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>